refactor(core): cut the wire compiler's dead patch mode, its duplicated action regex, and a pass-through re-export
- AttributeElement carried a fourth mode, "patch", for the edit dialect whose
  compiler was deleted; no caller has passed it since. Dropping it collapses
  three conditionals.
- ACTION_ATTR_PATTERN was written out in both the compiler and the printer,
  which are inverses of each other — one definition, imported.
- state.ts re-exported isWellFormedUtf16 from jcs.ts with a five-line comment
  about who reads it; the four readers now import jcs directly.
- The plan compiler re-spelled PLAN_DISPLAYS instead of checking membership,
  so a third display would have been declared and then refused.
